+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Registering Advanced Custom Fields integration hooks immediately on Unbabble register instead of waiting for `admin_init`.
 
+### Fixed
+
+- Fix missing action entry in $_POST on ajax call for new term.
+
 ## [0.5.7] - 2024-12-19
 
 ### Fixed