+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.8] - 2025-01-13
+
 ### Added
 
 - Setting Advanced Custom Fields fields as translatable automatically when they are registered.
@@ -542,7 +544,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First Release!
 
-[unreleased]: https://github.com/26b/unbabble/compare/0.5.7...HEAD
+[unreleased]: https://github.com/26b/unbabble/compare/0.5.8...HEAD
+[0.5.8]: https://github.com/26b/unbabble/compare/0.5.7...0.5.8
 [0.5.7]: https://github.com/26b/unbabble/compare/0.5.6...0.5.7
 [0.5.6]: https://github.com/26b/unbabble/compare/0.5.5...0.5.6
 [0.5.5]: https://github.com/26b/unbabble/compare/0.5.4...0.5.5